requirements.txt
build==1.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
idna==3.11
numpy==2.3.4
packaging==25.0
pygame-ce==2.5.6
pyproject_hooks==1.2.0
PyTMX==3.32
requests==2.32.5
setuptools==80.9.0
urllib3==2.5.0
wheel==0.45.1
